Keep input styling with the form, use react-hook-form
