+++ paywall/src/components/interface/MetadataForm.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
+import { useForm } from 'react-hook-form'
 import styled from 'styled-components'
 import { MetadataInput } from '../../unlockTypes'
-import { MetadataInput as Field } from './MetadataInput'
 import { ActionButton } from './buttons/ActionButton'
 
 interface Props {
@@ -10,18 +10,20 @@ interface Props {
 }
 
 export const MetadataForm = ({ fields, onSubmit }: Props) => {
-  const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
-    event.preventDefault()
-    // TODO: collect form data and submit as k/v pairs here
-    onSubmit({})
-  }
+  // We can also destructure the `errors` field here and use it for
+  // validation -- we'll have to consider how to handle the different
+  // kinds of errors so that we can show the right message
+  const { handleSubmit, register } = useForm()
 
   return (
-    <form onSubmit={handleSubmit}>
+    <form onSubmit={handleSubmit(onSubmit)}>
       <p>We need to collect some additional information for your purchase.</p>
       <FieldWrapper>
         {fields.map(({ name, type, required }) => (
-          <Field key={name} name={name} type={type} required={required} />
+          <Label required={required} key={name}>
+            <span>{name}</span>
+            <Input type={type} name={name} ref={register({ required })} />
+          </Label>
         ))}
       </FieldWrapper>
       <ActionButton type="submit">Continue</ActionButton>
@@ -33,4 +35,34 @@ export default MetadataForm
 
 const FieldWrapper = styled.div`
   margin-bottom: 32px;
+`
+
+const Input = styled.input`
+  height: 60px;
+  width: 100%;
+  background-color: var(--lightgrey);
+  border: medium none;
+  border-radius: 4px;
+  padding: 10px;
+  font-size: 16px;
+`
+
+interface LabelProps {
+  required: boolean
+}
+
+const Label = styled.label<LabelProps>`
+  & > span {
+    display: block;
+    text-transform: uppercase;
+    font-size: 10px;
+    color: var(--darkgrey);
+    margin-top: 16px;
+    margin-bottom: 5px;
+  }
+
+  & > span:after {
+    color: var(--red);
+    content: ${(props: LabelProps) => (props.required ? '" *"' : '')};
+  }
 `
